+++ src/nodes/r4s8cr.c
@@ -579,7 +579,7 @@ NODE_status_t R4S8CR_build_sigfox_ul_payload(NODE_ul_payload_update_t* ul_payloa
 		goto errors;
 	case NODE_SIGFOX_PAYLOAD_TYPE_DATA:
 		// Build registers list.
-		reg_list.addr_list = (uint32_t*) R4S8CR_REG_LIST_SIGFOX_PAYLOAD_ELECTRICAL;
+		reg_list.addr_list = (uint8_t*) R4S8CR_REG_LIST_SIGFOX_PAYLOAD_ELECTRICAL;
 		reg_list.size = sizeof(R4S8CR_REG_LIST_SIGFOX_PAYLOAD_ELECTRICAL);
 		// Reset registers.
 		status = XM_reset_registers(&reg_list, &node_reg);
